requirements.txt
flask
psycopg2-binary
python-dotenv
flask-cors
flask-talisman
flask-limiter
flask-jwt-extended
